requirements.txt
requests
tqdm
send2trash
pillow
mutagen
